Add functions for getting parameters of a specific flight
It turns out DFCX is incapable of something as simple as getting an element of a list by index, so I had to make a webhook function for it. This seems really stupid to me. Guess it's another complaint about DFCX to add to my growing list. Too bad DFCX wouldn't even be able to index that list properly.